feat: ai 분석 탭 목업 추가
