+++ lib/domain/util/duration_time.dart
@@ -1,26 +1,26 @@
+import 'dart:core';
+
+String twoDigits(int n) => n.toString().padLeft(2, "0");
+
 String formatDuration(Duration duration) {
-  String twoDigits(int n) => n.toString().padLeft(2, "0");
   String twoDigitMinutes = twoDigits(duration.inMinutes.remainder(60));
   String twoDigitSeconds = twoDigits(duration.inSeconds.remainder(60));
   return "${twoDigits(duration.inHours)}:$twoDigitMinutes:$twoDigitSeconds";
 }
 
 String formatDatetimeToHHMMSS(DateTime datetime) {
-  String twoDigits(int n) => n.toString().padLeft(2, "0");
   String twoDigitMinutes = twoDigits(datetime.minute.remainder(60));
   String twoDigitSeconds = twoDigits(datetime.minute.remainder(60));
   return "${twoDigits(datetime.hour)}:$twoDigitMinutes:$twoDigitSeconds";
 }
 
 String formatDatetimeToYYMMDD(DateTime datetime) {
-  String twoDigits(int n) => n.toString().padLeft(2, "0");
   String twoDigitDay = twoDigits(datetime.day);
   String twoDigitMonth = twoDigits(datetime.month);
   return "${datetime.year} $twoDigitMonth $twoDigitDay";
 }
 
 String formatDatetimeToAll(DateTime datetime) {
-  String twoDigits(int n) => n.toString().padLeft(2, "0");
   String twoDigitDay = twoDigits(datetime.day);
   String twoDigitMonth = twoDigits(datetime.month);
   String twoDigitMinutes = twoDigits(datetime.minute.remainder(60));
@@ -39,4 +39,11 @@ int formatTimeTextToSecond(String timeText) {
   } catch (_) {
     return 0;
   }
+}
+
+String formatTimestampToMMSS(int timestamp) {
+  Duration duration = Duration(milliseconds: timestamp);
+  String twoDigitMinute = twoDigits(duration.inMinutes);
+  String twoDigitSecond = twoDigits(duration.inSeconds % 60);
+  return "$twoDigitMinute:$twoDigitSecond";
 }
