+++ lib/presentation/ai_feedback/models/ai_score_state.dart
@@ -1,3 +1,4 @@
+import 'package:climb_balance/presentation/ai_feedback/models/ai_score_per_frame.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'ai_score_state.freezed.dart';
@@ -13,6 +14,16 @@ class AiScoreState with _$AiScoreState {
     @Default(0.0) double moment,
     @Default(0.0) double inertia,
   }) = _AiScoreState;
+
+  factory AiScoreState.fromPerFrame(AiScorePerFrame perFrameScore, int idx) {
+    return AiScoreState(
+      balance: perFrameScore.balance[idx] ?? 0,
+      accuracy: perFrameScore.accuracy[idx] ?? 0,
+      angle: perFrameScore.angle[idx] ?? 0,
+      moment: perFrameScore.moment[idx] ?? 0,
+      inertia: perFrameScore.inertia[idx] ?? 0,
+    );
+  }
 
   factory AiScoreState.fromJson(Map<String, dynamic> json) =>
       _$AiScoreStateFromJson(json);
