+++ pubspec.yaml
@@ -61,6 +61,7 @@ dependencies:
   go_router: ^5.1.1
   fluttertoast: ^8.0.9
   path_provider: ^2.0.11
+  carousel_slider: ^4.1.1
 dev_dependencies:
   flutter_test:
     sdk: flutter
